Fix parsing of highlight with small windows
If there are insufficient columns to list all attributes of a highlight
group on one line the output will split the output over multiple lines.
This happens not only interactively but also when we call highlight with
execute() or capture the output with redir. In such circumstances the
remaining attributes that extend past the newline are indented with
leading spaces such that they are visually aligned with the attributes
on the prior line.

This accounts for such a possibility and ensures that the full contents
of a highlight group is on a single line.