Cleanup variable naming in GetColors
Before we had both group_attributes and attributes. From the names one
would think they refer to differently scoped data, but in fact they
refer to the same data in different formats. Instead let us now refer to
the unparsed data as "values" and the parsed values as "attributes".

This commit also moves the initialization of the attributes dictionary.
Prior to this we initialized attributes as an empty dictionary and then
if the value was cleared (the first thing we check for) we'd change it
to a string, instead:

- If cleared declare the attributes as a string directly.
- If a link declare the attributes as a dictionary directly.
- Otherwise initialize as an empty dictionary so we may append() to it.