Make set function generic
If only given a group and no attributes clear, otherwise if attributes
are given clear and set said attributes. In the later case the clear is
needed to ensure any attributes that are already present but that we are
not setting are also removed.

This commit also includes some style changes.